Move Navigation bar related code to its own file
